+++ finish/src/main/java/io/openliberty/guides/inventory/InventoryResource.java
@@ -63,6 +63,7 @@ public class InventoryResource {
             example = "foo", 
             schema = @Schema(type = SchemaType.STRING)) 
         @PathParam("hostname") String hostname) {
+        // Get properties for host
         Properties props = manager.get(hostname);
         if (props == null) {
             return Response.status(Response.Status.NOT_FOUND)
@@ -70,6 +71,9 @@ public class InventoryResource {
                                 + "not be running on " + hostname)
                             .build();
         }
+
+        //Add to inventory to host
+        manager.add(hostname, props);
         return Response.ok(props).build();
     }
