+++ start/src/main/java/io/openliberty/guides/inventory/InventoryResource.java
@@ -42,6 +42,7 @@ public class InventoryResource {
     @Path("/{hostname}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response getPropertiesForHost(@PathParam("hostname") String hostname) {
+        // Get properties for host
         Properties props = manager.get(hostname);
         if (props == null) {
             return Response.status(Response.Status.NOT_FOUND)
@@ -49,6 +50,9 @@ public class InventoryResource {
                                 + "not be running on " + hostname)
                             .build();
         }
+
+        //Add to inventory to host
+        manager.add(hostname, props);
         return Response.ok(props).build();
     }
