+++ start/src/main/java/io/openliberty/guides/inventory/model/SystemData.java
@@ -1,0 +1,32 @@
+package io.openliberty.guides.inventory.model;
+
+import java.util.Properties;
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
+
+public class SystemData {
+
+    private final String hostname;
+    
+    private final Properties properties;
+
+    public SystemData(String hostname, Properties properties) {
+        this.hostname = hostname;
+        this.properties = properties;
+    }
+
+    public String getHostname() {
+        return hostname;
+    }
+
+    public Properties getProperties() {
+        return properties;
+    }
+
+    @Override
+    public boolean equals(Object host) {
+        if (host instanceof SystemData) {
+            return hostname.equals(((SystemData) host).getHostname());
+        }
+        return false;
+    }
+}
