+++ start/src/main/webapp/META-INF/openapi.yaml
@@ -14,6 +14,18 @@
 #       description: Server HTTP port.
 #       default: "9080"
 # paths:
+#   /inventory/systems:
+#     get:
+#       summary: List inventory contents.
+#       description: Returns the currently stored host:properties pairs in the inventory.
+#       operationId: listContents
+#       responses:
+#         200:
+#           description: host:properties pairs stored in the inventory.
+#           content:
+#             application/json:
+#               schema:
+#                 $ref: '#/components/schemas/InventoryList'
 #   /inventory/systems/{hostname}:
 #     get:
 #       summary: Get JVM system properties for particular host
@@ -40,30 +52,18 @@
 #             application/json:
 #               schema:
 #                 $ref: '#/components/schemas/Properties'
-#   /inventory/systems:
-#     get:
-#       summary: List inventory contents.
-#       description: Returns the currently stored host:properties pairs in the inventory.
-#       operationId: listContents
-#       responses:
-#         200:
-#           description: host:properties pairs stored in the inventory.
-#           content:
-#             application/json:
-#               schema:
-#                 $ref: '#/components/schemas/InventoryList'
 #   /inventory/properties:
 #     get:
-#       summary: Get JVM system properties.
-#       description: Returns the JVM system properties for the host running this service.
 #       operationId: getProperties
 #       responses:
-#         200:
-#           description: JVM system properties of the host running this service.
+#         default:
+#           description: default response
 #           content:
 #             application/json:
 #               schema:
-#                 $ref: '#/components/schemas/Properties'
+#                 type: object
+#                 additionalProperties:
+#                   type: string
 # components:
 #   schemas:
 #     InventoryList:
@@ -74,15 +74,11 @@
 #         systems:
 #           type: array
 #           items:
-#             $ref: '#/components/schemas/System'
+#             $ref: '#/components/schemas/SystemData'
 #         total:
 #           type: integer
 #       description: POJO that represents the inventory contents.
-#     Properties:
-#       type: object
-#       additionalProperties:
-#         type: string
-#     System:
+#     SystemData:
 #       required:
 #       - hostname
 #       - properties
@@ -95,3 +91,7 @@
 #           additionalProperties:
 #             type: string
 #       description: POJO that represents a single inventory entry.
+#     Properties:
+#       type: object
+#       additionalProperties:
+#         type: string
